moves fusion logic from fusion function to fuser
