+++ torch/csrc/jit/fusers/cuda/cuda_fuser.cpp
@@ -1,27 +1,25 @@
 #if defined USE_CUDA && !(defined _WIN32) && !(defined USE_ROCM)
 
-#include "torch/csrc/jit/fusers/cuda/cuda_fuser_interface.h"
 #include "torch/csrc/jit/fusers/cuda/cuda_fuser.h"
+#include "torch/csrc/jit/fusers/cuda/resource_strings.h"
 
 #include "torch/csrc/jit/ir.h"
-#include "torch/csrc/jit/code_template.h"
 #include "torch/csrc/jit/resource_guard.h"
 #include "torch/csrc/jit/constants.h"
-
 #include "torch/csrc/utils/disallow_copy.h"
 #include "torch/csrc/variable_tensor_functions.h"
 #include "torch/csrc/jit/assertions.h"
 
 #include "ATen/ATen.h"
-
+#include "ATen/DeviceGuard.h"
 #include "ATen/cuda/CUDAContext.h"
 #include "THC/THC.h"
 #include "THC/THCGenerator.hpp"
 #include "torch/csrc/cuda/cuda_check.h"
 
-#include <nvrtc.h>
-#include <cuda.h>
-#include <cuda_runtime.h>
+#include "nvrtc.h"
+#include "cuda.h"
+#include "cuda_runtime.h"
 
 #include <string>
 #include <algorithm>
@@ -33,8 +31,12 @@
 
 namespace torch { namespace jit { namespace cudafuser {
 
-// Note: there is only one CUDAFusionCompiler
-static CUDAFusionCompiler compiler;
+// Note: there is only one CUDAFuser
+static CUDAFuser fuser;
+
+CUDAFuser& getCUDAFuser() {
+  return fuser;
+}
 
 static const std::string check_exists_string =
   "which '${program}' > /dev/null";
@@ -45,217 +47,6 @@ static bool programExists(const std::string& program) {
   std::string cmd = format(check_exists_string, env);
   return 0 == system(cmd.c_str());
 }
-
-static const std::string so_template = "/tmp/pytorch_fuserXXXXXX.so";
-static const std::string cpp_template = "/tmp/pytorch_fuserXXXXXX.cpp";
-
-// NB: -march=native not supported on PPC64 g++.  It's a bit annoying
-// to do a configure-style test to decide whether or not the g++
-// actually supports it or not, so we heuristically use the host
-// compiler to predict if the runtime compiler supports the option we
-// want.  This probably won't work if you're cross-compiling.
-// NB: -march=native is disabled because it has caused problems where
-// compiler and assembler do not agree on what native instruction they
-// understand for AVX512. When we need better CPU performance this
-// optimization can be re-enabled by tracking down the platforms where
-// this error occurs and only selectively disabling it.
-static const std::string compile_string =
-  "\"${cxx}\" -O3 -g "
-#ifndef __PPC64__
-//  "-march=native "
-#endif
-  "-std=c++11 -fPIC ${fopenmp} -shared \"${cpp_file}\" -o \"${so_file}\" -lm";
-
-////////////////////////////////////////////////////////////////////////////////
-// Code generation
-
-/*with type_as not checking type of its input, a fusion group can have non-fp32 tensor as input.
-Correct code for this case is generated, however, nvrtc does not know how to handle int*_t integer types,
-so typedefs help it handle those cases*/
-
-auto type_declarations_template = CodeTemplate(R"(
-#if defined(__CUDACC_RTC__)
-typedef unsigned char uint8_t;
-typedef signed char int8_t;
-typedef short int  int16_t;
-typedef long long int int64_t;
-${HalfHeader}
-${RandHeader}
-#endif
-typedef ${IndexType} IndexType;
-template<typename T, size_t N>
-struct TensorInfo {
-  T * data;
-  IndexType sizes[N];
-  IndexType strides[N];
-};
-)");
-
-// We rewrite the code for philox RNG from curand as nvrtc couldn't resolve the
-// curand header correctly.
-constexpr auto rand_support_literal = R"(
-
-  class Philox {
-  public:
-    __device__ inline Philox(unsigned long long seed,
-                             unsigned long long subsequence,
-                             unsigned long long offset) {
-      key.x = (unsigned int)seed;
-      key.y = (unsigned int)(seed >> 32);
-      counter = make_uint4(0, 0, 0, 0);
-      counter.z = (unsigned int)(subsequence);
-      counter.w = (unsigned int)(subsequence >> 32);
-      STATE = 0;
-      incr_n(offset / 4);
-    }
-
-    __device__ inline unsigned long operator()() {
-      if(STATE == 0) {
-        uint4 counter_ = counter;
-        uint2 key_ = key;
-        for(int i = 0; i < 9; i++) {
-          counter_ = single_round(counter_, key_);
-          key_.x += (kPhilox10A); key_.y += (kPhilox10B);
-        }
-        output = single_round(counter_, key_);
-        incr();
-      }
-      unsigned long ret;
-      switch(STATE) {
-        case 0: ret = output.x; break;
-        case 1: ret = output.y; break;
-        case 2: ret = output.z; break;
-        case 3: ret = output.w; break;
-      }
-      STATE = (STATE + 1) % 4;
-      return ret;
-    }
-
-  private:
-    uint4 counter;
-    uint4 output;
-    uint2 key;
-    unsigned int STATE;
-    __device__ inline void incr_n(unsigned long long n) {
-      unsigned int nlo = (unsigned int)(n);
-      unsigned int nhi = (unsigned int)(n >> 32);
-      counter.x += nlo;
-      if (counter.x < nlo)
-        nhi++;
-      counter.y += nhi;
-      if (nhi <= counter.y)
-        return;
-      if (++counter.z)
-        return;
-      ++counter.w;
-    }
-    __device__ inline void incr() {
-      if (++counter.x)
-        return;
-      if (++counter.y)
-        return;
-      if (++counter.z)
-        return;
-      ++counter.w;
-    }
-    __device__ unsigned int mulhilo32(unsigned int a, unsigned int b,
-                                      unsigned int *result_high) {
-      *result_high = __umulhi(a, b);
-      return a*b;
-    }
-
-    __device__ inline uint4 single_round(uint4 ctr, uint2 key) {
-      unsigned int hi0;
-      unsigned int hi1;
-      unsigned int lo0 = mulhilo32(kPhiloxSA, ctr.x, &hi0);
-      unsigned int lo1 = mulhilo32(kPhiloxSB, ctr.z, &hi1);
-
-      uint4 ret = {hi1 ^ ctr.y ^ key.x, lo1, hi0 ^ ctr.w ^ key.y, lo0};
-      return ret;
-    }
-
-    static const unsigned long kPhilox10A = 0x9E3779B9;
-    static const unsigned long kPhilox10B = 0xBB67AE85;
-    static const unsigned long kPhiloxSA = 0xD2511F53;
-    static const unsigned long kPhiloxSB = 0xCD9E8D57;
-  };
-
-  // Inverse of 2^32.
-  #define M_RAN_INVM32 2.3283064e-10f
-  __device__  __inline__ float uniform(unsigned int x) {
-    return x * M_RAN_INVM32;
-  }
-)";
-
-constexpr auto rand_param = ",unsigned long long seed, unsigned long long offset";
-constexpr auto rand_init = R"(
-  int idx = blockIdx.x*blockDim.x + threadIdx.x;
-  Philox rnd(seed, idx, offset);
-)";
-
-auto cuda_compilation_unit_template = CodeTemplate(R"(
-${type_declarations}
-
-extern "C" __global__
-void ${kernelName}(IndexType totalElements, ${formals} ${RandParam}) {
-  ${RandInit}
-  for (IndexType linearIndex = blockIdx.x * blockDim.x + threadIdx.x;
-        linearIndex < totalElements;
-        linearIndex += gridDim.x * blockDim.x) {
-      // Convert `linearIndex` into an offset of tensor:
-      ${tensorOffsets}
-      // calculate the results
-      ${kernelBody}
-    }
-}
-)");
-
-// This snippet enables half support in the jit. Following the pattern for
-// reductions, fp16 input data is immediately upconverted to float
-// with __half2float(). All mathematical operations are done on float
-// values, and if needed the intermediate float representation is
-// converted to half with __float2half() when writing to a half tensor.
-constexpr auto half_support_literal  = R"(
-#define __HALF_TO_US(var) *(reinterpret_cast<unsigned short *>(&(var)))
-#define __HALF_TO_CUS(var) *(reinterpret_cast<const unsigned short *>(&(var)))
-#if defined(__cplusplus)
-  struct __align__(2) __half {
-    __host__ __device__ __half() { }
-
-  protected:
-    unsigned short __x;
-  };
-
-  /* All intrinsic functions are only available to nvcc compilers */
-  #if defined(__CUDACC__)
-    /* Definitions of intrinsics */
-    __device__ __half __float2half(const float f) {
-      __half val;
-      asm("{  cvt.rn.f16.f32 %0, %1;}\n" : "=h"(__HALF_TO_US(val)) : "f"(f));
-      return val;
-    }
-
-    __device__ float __half2float(const __half h) {
-      float val;
-      asm("{  cvt.f32.f16 %0, %1;}\n" : "=f"(val) : "h"(__HALF_TO_CUS(h)));
-      return val;
-    }
-  #endif /* defined(__CUDACC__) */
-#endif /* defined(__cplusplus) */
-#undef __HALF_TO_US
-#undef __HALF_TO_CUS
-
-typedef __half half;
-)";
-
-// curDimIndex = linearId % sizes[i]; // % sizes[i] is not needed for d == 0, because we already guard for numel outside the index calculation
-// offset += curDimIndex*strides[i]; // *strides[i] is optional if list_is_cont becaause strides.back() == 1
-// linearId /= sizes[i];
-auto dim_calc = CodeTemplate(R"(
-//printf("tensor ${tensor} sizes[${d}] = %d, strides[${d}] = %d\n", ${tensor}.sizes[${d}],${tensor}.strides[${d}]);
-size_t ${tensor}_dimIndex${d} = ${tensor}_linearIndex ${mod_sizes};
-${tensor}_offset += ${tensor}_dimIndex${d} ${times_stride};
-)");
 
 static void emitIndexingFor(
   std::ostream& out
@@ -525,83 +316,124 @@ std::pair<std::vector<ConcatDesc>, bool> emitCompilationUnit(
   return std::make_pair(std::move(concat_desc), has_random);
 }
 
-////////////////////////////////////////////////////////////////////////////////
-
-/*
-* CUDAFusionCompiler implementation.
-*/
-CUDAFusionCompiler::CUDAFusionCompiler() {
+CUDAFuser::CUDAFuser() {
   const char* cxx_env = getenv("CXX");
-  if(cxx_env != nullptr) config_.cxx = cxx_env;
-  if(!programExists(config_.cxx)) config_.cxx = "";
+  if (cxx_env != nullptr) config_.cxx = cxx_env;
+  if (!programExists(config_.cxx)) config_.cxx = "";
   const char* debug_env = getenv("PYTORCH_FUSION_DEBUG");
   config_.debug = debug_env && atoi(debug_env) != 0;
 }
 
-std::shared_ptr<CUDAFusionFunction> CUDAFusionCompiler::getOrCompile(
+static void checkCUDAVersion(const cudaDeviceProp& prop) {
+  if ((prop.major >= 6 && CUDA_VERSION < 8000) ||
+      (prop.major >= 7 && CUDA_VERSION < 9000)) {
+    std::stringstream err_string;
+    err_string << "PyTorch compiled with insufficient CUDA version for fusion: "
+         << CUDA_VERSION << " for the current GPU device " << prop.name
+         << " with device capability " << prop.major << "." << prop.minor;
+    throw std::runtime_error(err_string.str());
+  }
+}
+
+CUDAFusionFunction* CUDAFuser::constructFusionFunction(
+  std::string name
+, AnnotatedGraph& agraph) {
+  CUDAFusionFunction* func = new CUDAFusionFunction(name, agraph);
+
+  at::DeviceGuard device_guard(agraph.device);
+
+  CUDA_ASSERT(cudaGetDeviceProperties(&func->prop, agraph.device));
+  checkCUDAVersion(func->prop);
+
+  std::stringstream cu;
+  auto ret = emitCompilationUnit(cu, name, agraph, true);
+  func->concat_desc = std::move(ret.first);
+  func->has_random = ret.second;
+  func->compilation_unit = cu.str();
+  nvrtcProgram program;
+  NVRTC_ASSERT(nvrtcCreateProgram(
+    &program
+  , func->compilation_unit.c_str()
+  , NULL
+  , 0
+  , nullptr
+  , nullptr));
+
+  std::string compute = "--gpu-architecture=compute_" + std::to_string(func->prop.major) + std::to_string(func->prop.minor);
+  std::vector<const char *> args = {"--std=c++11", compute.c_str(), "-default-device"};
+  nvrtcResult result = nvrtcCompileProgram(program, args.size(), args.data());
+  
+  // Fails on error
+  // Note: special cases compilation error
+  if (result == NVRTC_ERROR_COMPILATION) {
+    size_t logsize;
+    nvrtcGetProgramLogSize(program, &logsize);
+    std::vector<char> log(logsize);
+    nvrtcGetProgramLog(program, log.data());
+    cu << log.data();
+    throw std::runtime_error(cu.str());
+  }
+  NVRTC_ASSERT(result);
+
+  ResourceGuard holdProgram([&] {
+    NVRTC_ASSERT(nvrtcDestroyProgram(&program));
+  });
+
+  size_t ptx_size;
+  NVRTC_ASSERT(nvrtcGetPTXSize(program, &ptx_size));
+  func->ptx.resize(ptx_size);
+  NVRTC_ASSERT(nvrtcGetPTX(program, func->ptx.data()));
+  
+  CU_ASSERT(cuModuleLoadData(&func->module, func->ptx.data()));
+  CU_ASSERT(cuModuleGetFunction(&func->function, func->module, func->name.c_str()));
+
+  CU_ASSERT(cuOccupancyMaxActiveBlocksPerMultiprocessor(
+    &func->maxBlocks
+  , func->function
+  , 128
+  , 0));
+  func->maxBlocks *= func->prop.multiProcessorCount;
+
+  return func;
+}
+
+std::shared_ptr<CUDAFusionFunction> CUDAFuser::getOrCompile(
   AnnotatedGraph& agraph) {
+  
+  JIT_ASSERT(agraph.device != kCPUDevice);
+
+  // Constructs key (for caching)
+  // TODO: abstract caching
   std::stringstream key;
   key << *agraph.graph << "\n";
   key << "device " << agraph.device << "\n";
-  for(auto & i : agraph.input_desc)
-    key << i << "\n";
-  for(auto & i : agraph.output_desc)
-    key << i << "\n";
+  for (auto& i : agraph.input_desc) key << i << "\n";
+  for (auto& i : agraph.output_desc) key << i << "\n";
   std::string key_ = key.str();
 
   auto it = cache.find(key_);
-  if (it == cache.end()) {
-    JIT_ASSERT(agraph.device != kCPUDevice);
+  if (it == cache.end()) {  
     std::string name = "kernel_" + std::to_string(cache.size());
-    CUDAFusionFunction* raw_func = new CUDAFusionFunction(name, agraph);
+    CUDAFusionFunction* raw_func = constructFusionFunction(name, agraph);
     it = cache.emplace(key_, std::shared_ptr<CUDAFusionFunction>(raw_func)).first;
   }
 
   return it->second;
 }
 
-std::shared_ptr<CUDAFusionFunction> CUDAFusionCompiler::getOrCompile(
+void CUDAFuser::debugLaunchGraph(
   Graph& graph
 , int device
 , at::ArrayRef<at::Tensor> inputs
 , at::ArrayRef<at::Tensor> outputs) {
   AnnotatedGraph agraph(graph, device);
-  for(auto& i : inputs) agraph.input_desc.emplace_back(i);
-  for(auto& i : outputs) agraph.output_desc.emplace_back(i);
-  return getOrCompile(agraph);
-}
-
-void CUDAFusionCompiler::debugLaunchGraph(
-  Graph& graph
-, int device
-, at::ArrayRef<at::Tensor> inputs
-, at::ArrayRef<at::Tensor> outputs) {
-  auto func = getOrCompile(graph, device, inputs, outputs);
+  for (auto& i : inputs) agraph.input_desc.emplace_back(i);
+  for (auto& i : outputs) agraph.output_desc.emplace_back(i);
+  auto func = getOrCompile(agraph);
   func->launch_with_tensors(inputs, outputs);
 }
 
 } // namespace cudafuser
-
-/*
-* Interface functions.
-*/
-std::shared_ptr<CompiledFusionFunction> getCUDAFusionFunction(Node* fusion_group) {
-  auto& graph = *fusion_group->g(attr::Subgraph);
-  cudafuser::AnnotatedGraph agraph(graph, fusion_group->i(attr::device));
-  
-  for(auto& input : graph.inputs()) {
-    auto t = input->type()->expect<TensorType>();
-    agraph.input_desc.emplace_back(t);
-  }
-
-  for(auto& output : graph.outputs()) {
-    auto t = output->type()->expect<TensorType>();
-    agraph.output_desc.emplace_back(t);
-  }
-
-  return cudafuser::compiler.getOrCompile(agraph);
-}
-
 } // namespace jit
 } // namespace torch
