+++ torch/csrc/jit/fusers/cuda/cuda_fuser.h
@@ -14,7 +14,6 @@
 #include "torch/csrc/jit/fusers/cuda/concat_desc.h"
 #include "torch/csrc/jit/fusers/cuda/cuda_fusion_function.h"
 #include "torch/csrc/jit/fusers/cuda/tensor_info.h"
-#include "torch/csrc/jit/fusers/cuda/temp_file.h"
 
 #include <string>
 #include <algorithm>
@@ -24,33 +23,14 @@
 
 namespace torch { namespace jit { namespace cudafuser {
 
-std::pair<std::vector<ConcatDesc>, bool> emitCompilationUnit(
-  std::ostream& out
-, const std::string& name
-, AnnotatedGraph& agraph
-, bool use_cuda);
-
-struct CUDAFusionCompilerConfig {
-  std::string cxx = "g++"; // compiler location
-  bool debug = false; // emit debugging information about fusions
-  bool openmp = true;
-};
-
-// caching compiler
-struct CUDAFusionCompiler {
-  TH_DISALLOW_COPY_AND_ASSIGN(CUDAFusionCompiler);
-  CUDAFusionCompiler();
+struct CUDAFuser {
+  TH_DISALLOW_COPY_AND_ASSIGN(CUDAFuser);
+  
+  CUDAFuser();
 
   // ignores types in graph, and uses specific contiguity annotations
   std::shared_ptr<CUDAFusionFunction> getOrCompile(
     AnnotatedGraph& agraph);
-
-  // uses inputs/outputs as examples to infer continuity, does not run the graph
-  std::shared_ptr<CUDAFusionFunction> getOrCompile(
-    Graph& graph
-  , int device
-  , at::ArrayRef<at::Tensor> inputs
-  , at::ArrayRef<at::Tensor> outputs);
 
   // debugging function that lets you do everything from compilation to execution
   // in one step.
@@ -63,11 +43,23 @@ struct CUDAFusionCompiler {
   , at::ArrayRef<at::Tensor> outputs);
   
 private:
-  CUDAFusionCompilerConfig config_;
+
+  CUDAFusionFunction* constructFusionFunction(
+    std::string name
+  , AnnotatedGraph& agraph);
+
+  struct CUDAFuserConfig {
+    std::string cxx = "g++"; // compiler location
+    bool debug = false; // emit debugging information about fusions
+  };
+
+  CUDAFuserConfig config_;
   std::unordered_map<
     std::string
   , std::shared_ptr<CUDAFusionFunction>> cache;
 };
+
+CUDAFuser& getCUDAFuser();
 
 
 } // namespace cudafuser
